+++ feature/profile-edit/impl/src/main/java/ru/yeahub/profile_edit/impl/ui/cropper/CropBottomSheet.kt
@@ -74,7 +74,7 @@ internal fun CropBottomSheet(
     val previewState = remember { CropPreviewState() }
 
     DisposableEffect(Unit) {
-        onDispose { previewState.recycle() }
+        onDispose { previewState.onDispose() }
     }
 
     val sheetState = rememberModalBottomSheetState(
